Move bundle related files to its own folder
