+++ composer.json
@@ -29,12 +29,14 @@
     },
     "autoload": {
         "psr-4": {
-            "SimpleThings\\EntityAudit\\": "src/SimpleThings/EntityAudit"
+            "SimpleThings\\EntityAudit\\": "src/SimpleThings/EntityAudit",
+            "SimpleThings\\EntityAuditBundle\\": "src/SimpleThings/EntityAuditBundle"
         }
     },
     "autoload-dev": {
         "psr-4": {
-            "SimpleThings\\EntityAudit\\Tests\\": "tests/SimpleThings/Tests/EntityAudit"
+            "SimpleThings\\EntityAudit\\Tests\\": "tests/SimpleThings/Tests/EntityAudit",
+            "SimpleThings\\Tests\\EntityAuditBundle\\": "tests/SimpleThings/Tests/EntityAudit"
         }
     },
     "extra": {
