+++ src/SimpleThings/EntityAuditBundle/DependencyInjection/Configuration.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace SimpleThings\EntityAudit\DependencyInjection;
+namespace SimpleThings\EntityAuditBundle\DependencyInjection;
 
 use Symfony\Component\Config\Definition\Builder\TreeBuilder;
 use Symfony\Component\Config\Definition\ConfigurationInterface;
