+++ src/SimpleThings/EntityAuditBundle/DependencyInjection/SimpleThingsEntityAuditExtension.php
@@ -21,7 +21,7 @@
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  */
 
-namespace SimpleThings\EntityAudit\DependencyInjection;
+namespace SimpleThings\EntityAuditBundle\DependencyInjection;
 
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
