+++ src/SimpleThings/EntityAuditBundle/Resources/views/Audit/compare.html.twig
