+++ src/SimpleThings/EntityAuditBundle/Resources/views/Audit/view_detail.html.twig
