+++ src/SimpleThings/EntityAuditBundle/Resources/views/Audit/view_entity.html.twig
