+++ src/SimpleThings/EntityAuditBundle/Resources/views/layout.html.twig
