+++ src/SimpleThings/EntityAuditBundle/SimpleThingsEntityAuditBundle.php
@@ -20,7 +20,7 @@
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  */
 
-namespace SimpleThings\EntityAudit;
+namespace SimpleThings\EntityAuditBundle;
 
 use Symfony\Component\HttpKernel\Bundle\Bundle;
