+++ tests/bootstrap.php
@@ -1,10 +1,11 @@
 <?php
 
-if (!($loader = @include __DIR__ . '/../vendor/autoload.php')) {
-    die(<<<'EOT'
+if (! ($loader = @include __DIR__ . '/../vendor/autoload.php')) {
+    echo <<<'EOT'
 You must set up the project dependencies, run the following commands:
 wget http://getcomposer.org/composer.phar
 php composer.phar install
-EOT
-    );
+EOT;
+
+    exit(1);
 }
