+++ tests/travis/pgsql.travis.xml
@@ -12,7 +12,7 @@
 
     <testsuites>
         <testsuite name="SimpleThings EntityAudit">
-            <directory>./../SimpleThings/Tests/EntityAudit</directory>
+            <directory>./../SimpleThings/Tests/</directory>
         </testsuite>
     </testsuites>
